simulation state refactor
